fix： naming

+++ app.js
@@ -18,23 +18,34 @@ app.get('/',(req,res)=>{
     res.render('home');
 });
 
-app.post('/uploadFile',(req,res)=>{
-    const file = req.files.file;
-    const fileName = req.body.fileName;
-    const filePath = 'files/'+fileName;
-
-    file.mv(filePath,async(err)=>{
-        if(err){
-            console.log("error : while uploading file");
-            return res.status(500).send(err);
+app.post('/upload-file',(req,res)=>{
+    try {
+        let file = null
+        for (let k in req.files) {
+            file = req.files[k]
         }
-        const fileHash = await addIpfsFile (fileName,filePath);
-        fs.unlink(filePath,(err)=>{
-            if(err) console.log(err);
-        })
-        res.render('upload',{fileName,fileHash});
-    })
+        if (file!==null) {
+            const fileName = file.name;
+            const filePath = 'files/'+fileName;
+            file.mv(filePath,async(err)=>{
+                if(err){
+                    console.log("error : while uploading file");
+                    return res.status(500).send(err);
+                }
+                const fileHash = await addIpfsFile (fileName,filePath);
+                fs.unlink(filePath,(err)=>{
+                    if(err) console.log(err);
+                })
+                res.render('upload',{fileName,fileHash});
+            })
+        } else {
+            res.status(500).send("no uploaded file");    
+        }
+    } catch (error) {
+        res.status(500).send(error);
+    }
 });
+
 const addIpfsFile = async (fileName,filePath)=>{
     const file = fs.readFileSync(filePath);
     const fileAdded = await ipfs.add({path: fileName,content:file});
